feat(refactor): move Runtime Arguments for Blueprints from StudioInstallation to ShowStyleBase
