+++ meteor/server/migration/0_18_0.ts
@@ -17,6 +17,7 @@ import * as _ from 'underscore'
 import {
 	PeripheralDevices,
 	PlayoutDeviceSettings,
+	PlayoutDeviceSettingsDevicePharos,
 	PlayoutDeviceSettingsDeviceHyperdeck,
 	PlayoutDeviceSettingsDevicePanasonicPTZ
 } from '../../lib/collections/PeripheralDevices'
